requirements.txt
fastapi==0.111.0
# pydantic and pydantic-core must be a matching pair (mixed conda/pip often breaks this).
pydantic==2.13.3
pydantic-core==2.46.3
uvicorn[standard]==0.29.0
ultralytics==8.2.0
python-multipart==0.0.9
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
# PyTorch 2.6+ defaults torch.load(weights_only=True); YOLO .pt checkpoints need full unpickle.
# Pin below 2.6 until Ultralytics fully supports safe loading (see torch.load release notes).
torch==2.5.1
torchvision==0.20.1